Moved ConstantObjectInspector PrimitiveObjectInspector and UnionObjectInspector
